Use custom TBD instead of linking with CoreBrightness directly
Apparently, the macOS SDK has a custom TBD file for CoreBrightness that restricts what symbols are available for linking to link with–in particular, only about a half dozen classes are included, even though many more are exported in the actual framework binary. If we create our own (local to this folder) TBD file that points to /System/Library/PrivateFrameworks/CoreBrightness.framework/CoreBrightness and "link" with that, instead of "linking" with the actual framework (which implicitly will cause the default TBD to be used), we can define the set of exported symbols we want to make available for use.